+++ src/components/Calendar/Calendar.js
@@ -1,0 +1,46 @@
+import React from 'react';
+import { withStyles } from '@material-ui/core/styles';
+
+import Grid from '@material-ui/core/Grid';
+import DateFnsUtils from '@date-io/date-fns';
+import {
+  MuiPickersUtilsProvider,
+  KeyboardDatePicker,
+} from '@material-ui/pickers';
+
+//стиль для календаря
+const CssGrid = withStyles({
+  root: {
+  	marginBottom: '20px'
+  },
+})(Grid);
+
+const Calendar = () => {
+  const [selectedDate, setSelectedDate] = React.useState(new Date('2020-09-08'));
+
+  const handleDateChange = (date) => {
+    setSelectedDate(date);
+  };
+
+  return (
+    <MuiPickersUtilsProvider utils={DateFnsUtils}>
+      <CssGrid container justify="space-around">
+        <KeyboardDatePicker
+          disableToolbar
+          variant="inline"
+          format="MM/dd/yyyy"
+          margin="normal"
+          id="date-picker-inline"
+          label="Date picker inline"
+          value={selectedDate}
+          onChange={handleDateChange}
+          KeyboardButtonProps={{
+            'aria-label': 'change date',
+          }}
+        />
+      </CssGrid>
+    </MuiPickersUtilsProvider>
+  )
+};
+
+export default Calendar;
